Split provider connect presentation
